\name{blima-package}
\alias{blima-package}
\alias{blima}
\docType{package}
\title{Package for analysis of Illumina microarrays bead level data. Package blima includes several bead level algorithms to support bead level and probe level analysis of the data. It utilizes its own method for background correction, performs variance stabilizing method on the bead level and bead level quantile normalization. It also provides some basic methods for differential expression testing.}
\description{Analysis of bead level data.}
\details{
\tabular{ll}{Package: \tab blima\cr
Type: \tab Package\cr
Title: \tab Package for analysis of Illumina microarrays bead level data. Package blima includes several bead level algorithms to support bead level and probe level analysis of the data. It utilizes its own method for background correction, performs variance stabilizing method on the bead level and bead level quantile normalization. It also provides some basic methods for differential expression testing.\cr
Version: \tab 0.99.29\cr
Date: \tab 2014-05-03\cr
Author: \tab Vojtech Kulvait\cr
Maintainer: \tab Vojtech Kulvait <kulvait@gmail.com>\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
Depends: \tab R(>= 3.0.0)\cr
Imports: \tab beadarray(>= 2.0.0), Biobase(>= 2.0.0), BiocGenerics, grDevices, stats, graphics\cr
Suggests: \tab xtable, blimaTestingData, BiocStyle, illuminaHumanv4.db, lumi\cr
URL: \tab https://bitbucket.org/kulvait/blima\cr
biocViews: \tab Microarray, Preprocessing, Normalization\cr}
}
\author{Vojtech Kulvait}

\keyword{ package }


