\name{checkIntegrityLogical}
\alias{checkIntegrityLogical}
\title{Internal function}
\description{Check integrity of the list of logical objects, internal.}
\usage{checkIntegrityLogical(xx, b, action = c("returnText", "warn", 
    "error"))}
\arguments{
  \item{xx}{List of logical objects compatible with a list b.}
  \item{b}{List of beadLevelData objects.}
  \item{action}{What type of action is required in case of invalid object structure. Either return text different from TRUE, warn or error.}
}



\author{Vojtech Kulvait}





