\name{checkIntegrityOfListOfBeadLevelDataObjects}
\alias{checkIntegrityOfListOfBeadLevelDataObjects}
\title{Internal function}
\description{Check integrity of the list of beadLevelData objects, internal.}
\usage{checkIntegrityOfListOfBeadLevelDataObjects(listb, action = c("returnText", 
    "warn", "error"))}
\arguments{
  \item{listb}{List of beadLevelData objects.}
  \item{action}{What type of action is required in case of invalid object structure. Either return text different from TRUE, warn or error.}
}



\author{Vojtěch Kulvait}





