\name{checkIntegrityOfSingleBeadLevelDataObject}
\alias{checkIntegrityOfSingleBeadLevelDataObject}
\title{Internal function}
\description{Check integrity of single beadLevelData object, internal.}
\usage{checkIntegrityOfSingleBeadLevelDataObject(b, action = c("returnText", 
    "warn", "error"))}
\arguments{
  \item{b}{beadLevelData object.}
  \item{action}{What type of action is required in case of invalid object structure. Either return text different from TRUE, warn or error.}
}



\author{Vojtěch Kulvait}





