\name{doAction}
\alias{doAction}
\title{Internal function}
\description{Performs action of certain type}
\usage{doAction(message, action = c("returnText", "warn", "error"))}
\arguments{
  \item{message}{Text message.}
  \item{action}{What type of action is required in case of invalid object structure. Either return text different from TRUE, warn or error.}
}



\author{Vojtech Kulvait}





