\name{getNextVector}
\alias{getNextVector}
\title{Support probe and beadl level testing.}
\description{Internal function supporting probe and beadl level testing.}
\usage{getNextVector(what, from, length)}
\arguments{
  \item{what}{Two column sorted matrix with probe values.}
  \item{from}{Index to start on}
  \item{length}{nrow(what)}
}



\author{Vojtěch Kulvait}





