\name{initMeanDistribution}
\alias{initMeanDistribution}
\title{initMeanDistribution}
\description{This is internal function not intended to direct use which initializes mean distribution.}
\usage{initMeanDistribution(srt, prvku)}
\arguments{
  \item{srt}{vector of sorted values}
  \item{prvku}{number of items in meanDistribution}
}



\author{Vojtech Kulvait}





