\name{insertColumn}
\alias{insertColumn}
\title{Internal function to support chipArrayStatistics}
\description{Internal}
\usage{insertColumn(matrix, column, name)}
\arguments{
  \item{matrix}{Object to insert column to}
  \item{column}{Column to insert}
  \item{name}{Name of column to assign.}
}



\author{Vojtech Kulvait}





