\name{interpolateSortedVector}
\alias{interpolateSortedVector}
\title{Interpolate sorted vector}
\description{Interpolates given sorted vector to the vector of different
length. For unsorted vectors do not guarantee any good properties.
Internal function.}
\usage{interpolateSortedVector(vector, newsize)}
\arguments{
  \item{vector}{Sorted vector to interpolate.}
  \item{newsize}{Size of the vector to produce.}
}



\author{Vojtech Kulvait}





