\name{numberOfDistributionElements}
\alias{numberOfDistributionElements}
\title{Internal}
\description{Internal function}
\usage{numberOfDistributionElements(b, normalizationMod = NULL, channelInclude = NULL)}
\arguments{
  \item{b}{Object beadLevelData from package beadarray or list of these objects}
  \item{normalizationMod}{NULL for normalization of all input b. Otherwise specifies logical vector of the length equals to the number of arrays in b or list of such vectors if b is a list of beadLevelData classes.}
  \item{channelInclude}{
}
}



\author{Vojtech Kulvait}





