\name{performXieCorrection}
\alias{performXieCorrection}
\title{INTERNAL FUNCTION Xie background correct.}
\description{INTERNAL This function is not intended for direct use.
Background correction according to non parametric estimator in
Xie, Yang, Xinlei Wang, and Michael Story. 
"Statistical Methods of Background Correction for Illumina BeadArray Data."
Bioinformatics 25, no. 6 (March 15, 2009): 751-57. doi:10.1093/bioinformatics/btp040. ###The method is applied on the bead level.}
\usage{performXieCorrection(value, alpha, mu, sigma)}
\arguments{
  \item{value}{
}
  \item{alpha}{
}
  \item{mu}{
}
  \item{sigma}{
}
}



\author{Vojtech Kulvait}





