\name{readToVector}
\alias{readToVector}
\title{Support doTTests function.}
\description{Internal function supporting doTTests function.}
\usage{readToVector(what, from, length, quality)}
\arguments{
  \item{what}{Item to read.}
  \item{from}{From index.}
  \item{length}{Length of vector.}
  \item{quality}{Column.}
}



\author{Vojtěch Kulvait}





