\name{singleCheckIntegrityLogicalVector}
\alias{singleCheckIntegrityLogicalVector}
\title{Internal function}
\description{Check integrity of the logical object, internal.}
\usage{singleCheckIntegrityLogicalVector(xx, b, action = c("returnText", 
    "warn", "error"))}
\arguments{
  \item{xx}{Logical object compatible with b.}
  \item{b}{Single beadLevelData object.}
  \item{action}{What type of action is required in case of invalid object structure. Either return text different from TRUE, warn or error.}
}



\author{Vojtěch Kulvait}





