\name{singleNumberOfDistributionElements}
\alias{singleNumberOfDistributionElements}
\title{Internal}
\description{Internal function}
\usage{singleNumberOfDistributionElements(b, normalizationMod = NULL, 
    channelInclude = NULL)}
\arguments{
  \item{b}{Object beadLevelData from package beadarray}
  \item{normalizationMod}{NULL for normalization of all input b. Otherwise specifies logical vector of the length equals to the number of arrays in b or list of such vectors if b is a list of beadLevelData classes.}
  \item{channelInclude}{
}
}



\author{Vojtěch Kulvait}





