\name{updateMeanDistribution}
\alias{updateMeanDistribution}
\title{updateMeanDistribution}
\description{This is internal function not intended to direct use. Updates mean distribution.}
\usage{updateMeanDistribution(meanDistribution, srt, arraysUsed)}
\arguments{
  \item{meanDistribution}{
}
  \item{srt}{vector of sorted values}
  \item{arraysUsed}{number of arrays allready used to create distribution}
}



\author{Vojtech Kulvait}





