\name{blima-package}
\alias{blima-package}
\alias{blima}
\docType{package}
\title{Package for the preprocessing and analysis of the Illumina microarrays on the detector (bead) level.}
\description{Package blima includes several algorithms for the preprocessing of Illumina microarray data. It focuses to the bead level analysis and provides novel approach to the quantile normalization of the vectors of unequal lengths. It provides variety of the methods for background correction including background subtraction, RMA like convolution and background outlier removal. It also implements variance stabilizing transformation on the bead level. There are also implemented methods for data summarization. It also provides the methods for performing T-tests on the detector (bead) level and on the probe level for differential expression testing.}
\details{
\tabular{ll}{Package: \tab blima\cr
Type: \tab Package\cr
Title: \tab Package for the preprocessing and analysis of the Illumina microarrays on the detector (bead) level.\cr
Version: \tab 1.1.03\cr
Date: \tab 2015-04-01\cr
Author: \tab Vojtech Kulvait\cr
Maintainer: \tab Vojtech Kulvait <kulvait@gmail.com>\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
Depends: \tab R(>= 3.0.0)\cr
Imports: \tab beadarray(>= 2.0.0), Biobase(>= 2.0.0), BiocGenerics, grDevices, stats, graphics\cr
Suggests: \tab xtable, blimaTestingData, BiocStyle, illuminaHumanv4.db, lumi\cr
URL: \tab https://bitbucket.org/kulvait/blima\cr
biocViews: \tab Microarray, Preprocessing, Normalization\cr}
}
\author{Vojtech Kulvait}

\keyword{ package }


