\name{blima-package}
\alias{blima-package}
\alias{blima}
\docType{package}
\title{Package for the preprocessing and analysis of the Illumina microarrays on the detector (bead) level.}
\description{Package blima includes several algorithms for the preprocessing of Illumina microarray data. It focuses to the bead level analysis and provides novel approach to the quantile normalization of the vectors of unequal lengths. It provides variety of the methods for background correction including background subtraction, RMA like convolution and background outlier removal. It also implements variance stabilizing transformation on the bead level. There are also implemented methods for data summarization. It provides the methods for performing T-tests on the detector (bead) level and on the probe level for differential expression testing.}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{blima}
\packageIndices{blima}
}
\author{Vojtěch Kulvait
Maintainer: \packageMaintainer{blima}
}
\keyword{ package }


