\name{interpolateSortedVector}
\alias{interpolateSortedVector}
\title{Interpolate sorted vector}
\description{Interpolates given sorted vector to the vector of different
length. It does not sort input vector thus for unsorted vectors do not guarantee functionality.
Internal function.}
\usage{interpolateSortedVector(vector, newSize)}
\arguments{
  \item{vector}{Sorted vector to interpolate.}
  \item{newSize}{Size of the vector to produce.}
}



\author{Vojtěch Kulvait}





